Finished return_item (not tested), added several functions in description.
The completion of return_item led to the creation of several functions. is_late, return_hours, and wipe_trans_history are all byproducts of return_item. At this point, the rabbit hole is getting deeper and I haven't been able to test return_item, so it may not be in working order. I also fixed an error in view_trans_or_inventory where an error was thrown because a datetime wasn't being converted to a string.