Added several functions, finished rent, fixed minor bugs.
I added calculate_return_date and choice_num_to_item. Also finished rent and it is in working order (as far as I know). There was a TypeError in view_inventory_or_trans_history with trying to format a datetime object that was actually a string, so I fixed that by added a branch. I changed all the IDs to lowercase for ease of access more than anything. In addition, I imported timedelta for some of the time difference type functions.