Changed some formats to namedtuples, fixed various bugs.
With my current use of mypy, types are becoming more and more important. This being said, a couple of the data structures I was using were inferior and caused mypy to get upset. So I changed those over to namedtuples instead. Now, they're much more efficient and easy to use.