Moved some things, changed type contracts to declarations, made view prettier.
I moved calculate_return_date down just for aesthetic's sake. I also moved a dictionary from return_item to choose_trans. The biggest addition on this commit are the changing of the type contracts to mypy declarations. A structure in is_late was sketchy, so I changed it to a namedtuple and changed other code involved. The view_inventory... function was made prettier by adding separating lines and more newlines.